# fix: `cargo test --profile release`